backend/requirements.txt
fastapi
uvicorn
python-socketio
supabase
httpx
python-dotenv
pydantic
